added dark theme, refactored structure
