+++ frontend/package.json
@@ -12,8 +12,10 @@
     "axios": "^1.7.9",
     "next": "15.1.6",
     "next-auth": "^4.24.11",
+    "next-themes": "^0.4.4",
     "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "react-dom": "^19.0.0",
+    "react-icons": "^5.4.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
