+++ frontend/src/components/theme/providers.tsx
@@ -1,0 +1,11 @@
+"use client";
+import { ThemeProvider } from "next-themes";
+import { ReactNode } from "react";
+
+export default function Providers({ children }: { children: ReactNode }) {
+  return (
+    <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
+      {children}
+    </ThemeProvider>
+  );
+}
