+++ frontend/tailwind.config.ts
@@ -15,4 +15,5 @@ export default {
     },
   },
   plugins: [],
+  darkMode: "class",
 } satisfies Config;
